refactor: move extractErrorMessage from @utils/core to @common/errors
`extractErrorMessage` is an error utility that was living in
`src/utils/core/stringCore.ts` — a string formatting module — and
exported from the `@utils/core` barrel. Its natural home is alongside
`toErrorMessage` and `ensureError` in `@common/errors`.

- Move implementation into `src/common/errors/errorMessage.ts`
  (imports `isNonEmptyString` from `@utils/core` to avoid duplication)
- Export from the `@common/errors` public barrel
- Remove from `src/utils/core/stringCore.ts` and the `@utils/core` barrel
- Clean up the now-stale "Error Utility Guide" section in `stringCore.ts`
- Update internal caller `sdkError/providerErrorFormat.ts` to import
  directly from `'../errorMessage'`
- Update external caller `packages/cli/src/runtime/doctor.ts` to use
  `@common/errors`

No functional change — implementation is identical; only ownership is
clarified. All error extraction utilities now live in one module.
