+++ src/common/errors/sdkError/providerErrorFormat.ts
@@ -3,7 +3,7 @@ import {
   type ErrorLogData,
   type ProviderError,
 } from '@shared/schemas';
-import { extractErrorMessage } from '@utils/core';
+import { extractErrorMessage } from '../errorMessage';
 
 import { toErrorMessage } from '../errorMessage';
 import { isDiskFullError } from '../errorPredicates';
