+++ src/utils/core/index.ts
@@ -8,7 +8,6 @@
 export {
   isNonEmptyString,
   isString,
-  extractErrorMessage,
   formatCompactDuration,
   formatCompactTokenCount,
   formatDuration,
